requirements.txt
streamlit
pillow
requests
pandas
python-dotenv

